List indices need to be integers
In Python 2 `math`floor()` returns a float.
In Python 3 it returns an integer.

+++ demo/demo_advection_1D.py
@@ -30,8 +30,8 @@ from pyns.physical       import properties
 
 sens = -1, +1
 test = X, Y, Z
-SENS = sens[ floor( random() * len(sens) ) ]
-TEST = test[ floor( random() * len(test) ) ]
+SENS = sens[int(floor( random() * len(sens)))]
+TEST = test[int(floor( random() * len(test)))]
 MIN = 10
 MAX = 20
